requirements.txt
langchain-openai==0.0.2
openai==1.7.1
antimatter~=0.1.3
